Removed Integration test from native client and make sure rest client integration test runs with docker image.
