+++ elasticsearch/rest/src/test/resources/Dockerfile
